requirements.txt
flask
flask-cors
tensorflow
librosa
numpy
pickle5
SpeechRecognition
pydub
google-genai
python-dotenv